refactor: 디렉토리 구조 변경 및 ESM으로 변경
