+++ main.html
@@ -217,7 +217,5 @@
         </section>
     </div>
     </body>
-    <script src="/script.js"></script>
-    <script src="/bringDate.js"></script>
-    <script src="/rollingBanner.js"></script>
+    <script type="module" src="./app.js"></script>
 </html>
